+++ README.md
@@ -7,7 +7,6 @@
   <a href="https://thementorship.tw" target="blank"><img src="https://img.shields.io/website?url=https%3A%2F%2Fthementorship.tw&logo=github&style=flat-square" /></a>
 </p>
 
-
 <img src="nextjs-app/public/images/header-logo.png" alt="thementorship-logo" width="182" height="40" />
 
 ## About The Mentorship Program
@@ -22,7 +21,6 @@ We use Next.js for the frontend and Sanity for the headless CMS.
 
 This project is forked from [sanity-template-nextjs-clean](https://github.com/sanity-io/sanity-template-nextjs-clean), A clean Next.js template with a native authoring experience.
 
-
 ### Key Features
 
 - **Next.js 15 for Performance:** Leverage the power of Next.js 15 App Router for blazing-fast performance and SEO-friendly static sites.
@@ -33,7 +31,6 @@ This project is forked from [sanity-template-nextjs-clean](https://github.com/sa
 - **AI-powered Media Support:** Auto-generate alt text with [Sanity AI Assist](https://www.sanity.io/ai-assist).
 - **On-demand Publishing:** No waiting for rebuilds—new content is live instantly with Incremental Static Revalidation.
 - **Easy Media Management:** [Integrated Unsplash support](https://www.sanity.io/plugins/sanity-plugin-asset-source-unsplash) for seamless media handling.
-
 
 ## Repository Structure
 
@@ -52,16 +49,16 @@ For the development guide for studio, please see [studio/README.md](studio/READM
 Join our community and stay updated:
 
 <a href="https://www.facebook.com/thementorshiptaiwan" target="_blank">
-  <img src="nextjs-app/public/images/facebook-logo.svg" alt="facebook-logo" width="32" height="32" />
+  <img src="nextjs-app/public/images/social-media/facebook-filled.svg" alt="facebook-logo" width="32" height="32" />
 </a>
 <a href="https://www.instagram.com/thementorshiptaiwan/" target="_blank">
-  <img src="nextjs-app/public/images/instagram-logo.svg" alt="instagram-logo" width="32" height="32" />
+  <img src="nextjs-app/public/images/social-media/instagram-filled.svg" alt="instagram-logo" width="32" height="32" />
 </a>
 <a href="https://www.linkedin.com/company/the-mentorship-taiwan/posts" target="_blank">
-  <img src="nextjs-app/public/images/linkedin-logo.svg" alt="linkedin-logo" width="32" height="32" />
+  <img src="nextjs-app/public/images/social-media/linkedin-filled.svg" alt="linkedin-logo" width="32" height="32" />
 </a>
 <a href="https://thementorshiptaiwan.medium.com/" target="_blank">
-  <img src="nextjs-app/public/images/medium-logo.svg" alt="medium-logo" width="32" height="32" />
+  <img src="nextjs-app/public/images/social-media/medium-filled.svg" alt="medium-logo" width="32" height="32" />
 </a>
 
 ## Deployment Partner
